Use %zu for size_t variables in (some) format strings

+++ src/XrdMacaroons/XrdMacaroonsHandler.cc
@@ -573,7 +573,7 @@ Handler::GenerateMacaroonResponse(XrdHttpExtReq &req, const std::string &resourc
     std::vector<char> macaroon_resp; macaroon_resp.resize(size_hint);
     if (macaroon_serialize(mac_with_date, &macaroon_resp[0], size_hint, &mac_err))
     {
-        printf("Returned macaroon_serialize code: %lu\n", (unsigned long)size_hint);
+        printf("Returned macaroon_serialize code: %zu\n", size_hint);
         return req.SendSimpleResp(500, NULL, NULL, "Internal error serializing macaroon", 0);
     }
     macaroon_destroy(mac_with_date);
